Delete abstract class and rename a file
