+++ bootstrap.py
@@ -1,7 +1,7 @@
 from shared.gast_to_code.converter_registry import ConverterRegistry
 from python.gast_to_code.py_gast_to_code_converter import PyGastToCodeConverter
 from javascript.gast_to_code.js_gast_to_code_converter import JsGastToCodeConverter
-from java.gast_to_code.gast_to_code_java import JavaGastToCodeConverter
+from java.gast_to_code.java_gast_to_code_converter import JavaGastToCodeConverter
 from bash.gast_to_code.bash_gast_to_code_converter import BashGastToCodeConverter
