--- shared/gast_to_code/abstract_gast_to_code_converter.py
@@ -1,28 +1,0 @@
-from abc import ABC, abstractmethod
-
-
-class AbstractGastToCodeConverter(ABC):
-
-    @abstractmethod
-    def handle_log_statement(gast):
-        pass
-
-    @abstractmethod
-    def handle_var_assign(gast):
-        pass
-
-    @abstractmethod
-    def handle_bool(gast):
-        pass
-
-    @abstractmethod
-    def handle_if(gast):
-        pass
-
-    @abstractmethod
-    def handle_functions(gast):
-        pass
-
-    @abstractmethod
-    def handle_root(gast):
-        pass
